Remove nix docker config
Cannot find anything online for how to build off of existing images of other distros. This is something that can very easily be done in docker-compose, yet seems almost impossible in NixOS. If it's possible, someone should have documented it. For now I go back to using docker compose files for the freedom.