Make workstation's system backup entirety of root
Originally only /home was backed up, however containers and other useful data is stored in /var/lib, and some useful data is stored in /etc. I decided at this point that it would be best to just include the entirety of root and exclude anything that is massive and unneeded.