Fix Picom for selkies gstreamer containers
Earlier the Picom default backend changed to glx because its default xrender performance was absolutely terrible (and rather than abandon Picom or copy paste glx in all my systems, I decided it should be the new default). Selkies Gstreamer uses VirtualGL for allowing isolated desktops to share the GPU, but when Picom is run with VirtualGL with the glx backend, it crashes at startup. The default container desktop config sets Picom's backend to xrender so that Picom will continue to work correctly.