Update Selkies Gstreamer, add pulse requirement
pulseaudio.service was added as a selkies-gstreamer requirement because the web page will be stuck at 'Waiting for stream' until pulseaudio is running, if pulseaudio is not running then it leaves the end user confused as to why it's never loading. Now if pulseaudio isn't running, the web page won't work and it will make the user aware that something is wrong with the container rather than whatever else they may suspect.