Refactor workstation flatpak, fix gamescope crashing
Garry's Mod was crashing on startup when ran with gamescope, which wasn't a problem with Fallout 4. I figured the issue must've been a version mismatch between gamescope flatpak and proton flatpak, so I downgraded proton flatpak to the 23.08 commit, which lines up with 23.08 mangohud and gamescope. It's all working fine now.