Separate flatpak steam and its gamescope
There is a Steam Overlay issue when you run Steam directly with Gamescope. What happens is the Overlay won't work whatsoever, and that includes Steam screenshots. After lots of research, I determined that starting Gamescope and then attaching processes such as Steam to it at a later time, is really the only current way to run Steam and its overlay in Gamescope.